Browse node ID value can exceed 4294967295

+++ sql/test/browse_node/create.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `browse_node` (
-  `browse_node_id` int(10) unsigned NOT NULL,
+  `browse_node_id` BIGINT(10) unsigned NOT NULL,
   `name` varchar(255) NOT NULL,
   PRIMARY KEY (`browse_node_id`)
 ) ENGINE=InnoDB CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
